+++ src/typecheck.c
@@ -1664,7 +1664,7 @@ PUREFUNC bool is_constant(env_t *env, ast_t *ast) {
     case TextLiteral: {
         Text_t literal = Match(ast, TextLiteral)->text;
         TextIter_t state = NEW_TEXT_ITER_STATE(literal);
-        for (int64_t i = 0; i < literal.length; i++) {
+        for (int64_t i = 0; i < (int64_t)literal.length; i++) {
             int32_t g = Text$get_grapheme_fast(&state, i);
             if (g < 0 || g > 127 || !isascii(g)) return false;
         }
